fix(BACKEND_URL): separate backend url by target platform
