+++ .gitignore
@@ -47,3 +47,5 @@ main.js.map
 
 .idea
 storybook-static
+favicon.ico
+index.html
