+++ app/backendUrl.js
@@ -1,0 +1,1 @@
+export { BACKEND_URL } from './platformSpecific/backendUrl.TARGET_PLATFORM'
