+++ app/components/modals/ExportDataset.tsx
@@ -12,7 +12,7 @@ import Error from './Error'
 import CommitDetails from '../CommitDetails'
 import RadioInput from '../form/RadioInput'
 import { refStringFromQriRef } from '../../models/qriRef'
-import { BACKEND_URL } from '../../constants'
+import { BACKEND_URL } from '../../backendUrl'
 import { VersionInfo } from '../../models/store'
 import { ExportSubmitButton } from './ExportSubmitButton'
 import Button from '../chrome/Button'
