+++ app/constants.js
@@ -4,7 +4,7 @@
 // if we are targeting the remote build & we are in prod mode, we want the
 // backend url to be the url that the user is viewing the webapp from
 // otherwise, we are trying to talk to the local qri backend
-const BACKEND_URL = process.env.REMOTE && process.env.NODE_ENV === 'production' ? window.location.origin : 'http://localhost:2503'
+
 const CRASH_REPORTER_URL = 'https://crashreports.qri.io/desktop'
 const DISCORD_URL = 'https://discordapp.com/invite/thkJHKj'
 const GITHUB_ORG_URL = 'https://github.com/qri-io'
@@ -17,7 +17,6 @@ const WEBSOCKETS_PROTOCOL = 'qri-websocket'
 const DEFAULT_POLL_INTERVAL = 3000
 
 module.exports = {
-  BACKEND_URL,
   CRASH_REPORTER_URL,
   DEFAULT_POLL_INTERVAL,
   DISCORD_URL,
