+++ app/platformSpecific/backendUrl.electron.js
@@ -1,0 +1,5 @@
+const BACKEND_URL = 'http://localhost:2503'
+
+module.exports = {
+  BACKEND_URL
+}
