+++ app/platformSpecific/backendUrl.web.js
@@ -1,0 +1,3 @@
+export const BACKEND_URL = process.env.NODE_ENV === 'production' && __BUILD__.REMOTE ? window.location.origin : 'http://localhost:2503'
+
+export default BACKEND_URL
